requirements.txt
# common
netifaces
paho-mqtt
pyserial
python-can == 4.3
pytest
python-dotenv
# dronecan related
pydronecan
raccoonlab-tools

# cyphal related
pycyphal
yakut

pytest-asyncio
# bot
pyTelegramBotAPI
aiogram<3.0

#rpi
RPi.GPIO
